Move measureElement HOC
